Explicitly cast int to uint16_t to satisfy MSVC compiler warnings

+++ samples/mqtt_pub_sub/main.cpp
@@ -129,7 +129,7 @@ int main(int argc, char *argv[])
 
         if (s_cmdOptionExists(argv, argv + argc, "--proxy_port"))
         {
-            proxyPort = atoi(s_getCmdOption(argv, argv + argc, "--proxy_port"));
+            proxyPort = static_cast<uint16_t>(atoi(s_getCmdOption(argv, argv + argc, "--proxy_port")));
         }
     }
